modificacion de AddClient
