+++ src/Login/Login.css
@@ -6,6 +6,6 @@
     }
 
     .cardSize{
-      width: 20rem;
+      width: 16rem;
     }
   }
